mlops_project/web_service_mlflow_visualiztion/requirements.txt
flask
gunicorn
pandas
boto3
python-dotenv
mlflow 
autogluon.tabular[lightgbm, xgboost, catboost]
prometheus_client